change router and css style
