+++ package.json
@@ -13,7 +13,6 @@
     "next": "12.3.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-router-dom": "^6.4.2",
     "styled-components": "^5.3.6"
   },
   "devDependencies": {
